splitted rve module into rve_utils and rve_utils_periodic modules
